requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-dotenv>=1.0
python-multipart

langchain>=0.2
langchain-community>=0.2
langchain-text-splitters>=0.2
langchain-huggingface>=0.1.0

huggingface_hub>=0.24
sentence-transformers>=3.0
faiss-cpu>=1.8.0

pypdf>=4.0
